+++ packages/prop-order/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 3.0.0
+
+ - Migrate to `stylelint` `v16.0.0`
+
 ## 2.0.16
 
 - Add new properties: